twibd: refactor TCP, UNIX frontends into unified SocketFrontend
