+++ twib/twibd/CMakeLists.txt
@@ -1,11 +1,7 @@
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
-set(SOURCE Twibd.cpp USBBackend.cpp Messages.cpp LocalClient.cpp msgpack_show.cpp TCPFrontend.cpp)
-if(NOT WIN32)
-	set(SOURCE ${SOURCE} UNIXFrontend.cpp)
-endif()
-
+set(SOURCE Twibd.cpp USBBackend.cpp Messages.cpp LocalClient.cpp msgpack_show.cpp SocketFrontend.cpp)
 add_executable(twibd ${SOURCE})
 
 target_link_libraries(twibd twib-log twib-buffer)
