+++ twib/twibd/Twibd.hpp
@@ -13,11 +13,6 @@
 
 #include "USBBackend.hpp"
 
-#include "TCPFrontend.hpp"
-#ifndef _WIN32
-#include "UNIXFrontend.hpp"
-#endif
-
 #include "Logger.hpp"
 #include "Messages.hpp"
 #include "Device.hpp"
@@ -43,10 +38,6 @@ class Twibd {
 	
  private:
 	backend::USBBackend usb;
-	frontend::TCPFrontend tcp;
-#ifndef _WIN32
-	frontend::UNIXFrontend unix;
-#endif
 
 	moodycamel::BlockingConcurrentQueue<std::variant<Request, Response>> dispatch_queue;
